requirements.txt
playwright==1.47.0
openpyxl==3.1.5
Pillow==10.4.0
